+++ CHANGELOG.md
@@ -8,6 +8,7 @@ All notable changes to this project will be documented in this file.
 - Added disabling `Tickets` button in client menu. To enable this tweak set `HideTicketsFromPlayers` to true in sandbox options.
 - Added storage for admin powers. The values of the checkboxes in the admin powers window are saved and powered user does not need to change options after rejoining the server. Also adds "Show admin tag" option to Admin Powers checkboxes. To enable this tweak set `SaveAdminPower` to true in sandbox options.
 - Added Safehouse to spawn locations and moved it as first location. To enable this tweak set `AddSafehouseToSpawnLocations` to true in sandbox options.
+- Added allowing for GM (and higher levelled users) to add anyone to safehouse without limits. To enable this tweak set `AdminsFreeAddToSafehouse` to true in sandbox options.
 
 ## [v0.2.0] - 2023-03-29
 ### Added